+++ Libplanet.Tests/Net/PeerTest.cs
@@ -33,7 +33,6 @@ namespace Libplanet.Tests.Net
                         0x1a, 0x3d, 0x3c, 0x76, 0xdb,
                     }),
                     new DnsEndPoint("0.0.0.0", 1234),
-                    default(AppProtocolVersion),
                     IPAddress.IPv6Loopback),
             };
             yield return new object[]
@@ -49,8 +48,7 @@ namespace Libplanet.Tests.Net
                         0x32, 0xfd, 0xa7, 0xdd, 0xc4, 0x4a, 0x16, 0x95, 0xe5, 0xce,
                         0x1a, 0x3d, 0x3c, 0x76, 0xdb,
                     }),
-                    new DnsEndPoint("0.0.0.0", 1234),
-                    ver),
+                    new DnsEndPoint("0.0.0.0", 1234)),
             };
             yield return new object[]
             {
@@ -65,8 +63,7 @@ namespace Libplanet.Tests.Net
                         0x32, 0xfd, 0xa7, 0xdd, 0xc4, 0x4a, 0x16, 0x95, 0xe5, 0xce,
                         0x1a, 0x3d, 0x3c, 0x76, 0xdb,
                     }),
-                    new DnsEndPoint("0.0.0.0", 1234),
-                    ver),
+                    new DnsEndPoint("0.0.0.0", 1234)),
             };
             yield return new object[]
             {
@@ -80,8 +77,7 @@ namespace Libplanet.Tests.Net
                         0xac, 0x2e, 0xf6, 0xc6, 0xee, 0x05, 0xdb, 0x06, 0xa9, 0x45,
                         0x32, 0xfd, 0xa7, 0xdd, 0xc4, 0x4a, 0x16, 0x95, 0xe5, 0xce,
                         0x1a, 0x3d, 0x3c, 0x76, 0xdb,
-                    }),
-                    ver2),
+                    })),
             };
         }