+++ Libplanet/Net/Messages/Message.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
 using System.Runtime.Serialization.Formatters.Binary;
@@ -11,6 +12,8 @@ namespace Libplanet.Net.Messages
 {
     internal abstract class Message
     {
+        public const int CommonFrames = 4;
+
         internal enum MessageType : byte
         {
             /// <summary>
@@ -105,9 +108,41 @@ namespace Libplanet.Net.Messages
             /// Contains the delta states of the requested block.
             /// </summary>
             BlockStates = 0x23,
+
+            /// <summary>
+            /// A reply to any messages with different <see cref="AppProtocolVersion"/>.
+            /// Contains the expected and actual <see cref="AppProtocolVersion"/>
+            /// value of the message.
+            /// </summary>
+            DifferentVersion = 0x30,
+        }
+
+        private enum MessageFrame
+        {
+            /// <summary>
+            /// Frame containing <see cref="AppProtocolVersion"/>.
+            /// </summary>
+            Version = 0,
+
+            /// <summary>
+            /// Frame containing <see cref="MessageType"/>.
+            /// </summary>
+            Type = 1,
+
+            /// <summary>
+            /// Frame containing the sender <see cref="Peer"/> of the <see cref="Message"/>.
+            /// </summary>
+            Peer = 2,
+
+            /// <summary>
+            /// Frame containing signature of the <see cref="Message"/>.
+            /// </summary>
+            Sign = 3,
         }
 
         public byte[] Identity { get; set; }
+
+        public AppProtocolVersion Version { get; set; }
 
         public Peer Remote { get; set; }
 
@@ -115,21 +150,54 @@ namespace Libplanet.Net.Messages
 
         protected abstract IEnumerable<NetMQFrame> DataFrames { get; }
 
-        public static Message Parse(NetMQMessage raw, bool reply)
+        public static Message Parse(
+            NetMQMessage raw,
+            bool reply,
+            AppProtocolVersion localVersion,
+            IImmutableSet<PublicKey> trustedAppProtocolVersionSigners,
+            DifferentAppProtocolVersionEncountered differentAppProtocolVersionEncountered)
         {
             if (raw.FrameCount == 0)
             {
                 throw new ArgumentException("Can't parse empty NetMQMessage.");
             }
 
-            // (reply == true)  [type, sign, peer, frames...]
-            // (reply == false) [identity, type, sign, peer, frames...]
-            int headerCount = reply ? 3 : 4;
-            var rawType = (MessageType)raw[headerCount - 3].ConvertToInt32();
-            var peer = raw[headerCount - 2].ToByteArray();
-            byte[] signature = raw[headerCount - 1].ToByteArray();
+            // (reply == true)  [version, type, peer, sign, frames...]
+            // (reply == false) [identity, version, type, peer, sign, frames...]
+            NetMQFrame[] remains = reply ? raw.ToArray() : raw.Skip(1).ToArray();
 
-            NetMQFrame[] body = raw.Skip(headerCount).ToArray();
+            var versionToken = remains[(int)MessageFrame.Version].ConvertToString();
+
+            AppProtocolVersion remoteVersion = AppProtocolVersion.FromToken(versionToken);
+            Peer remotePeer = null;
+            try
+            {
+                remotePeer = DeserializePeer(remains[(int)MessageFrame.Peer].ToByteArray());
+            }
+            catch (Exception)
+            {
+                // If failed to find out remotePeer, leave it null.
+            }
+
+            if (!IsAppProtocolVersionValid(
+                remotePeer,
+                localVersion,
+                remoteVersion,
+                trustedAppProtocolVersionSigners,
+                differentAppProtocolVersionEncountered))
+            {
+                throw new DifferentAppProtocolVersionException(
+                    "Received message's version is not valid.",
+                    reply ? null : raw[0].Buffer.ToArray(),
+                    localVersion,
+                    remoteVersion);
+            }
+
+            var rawType = (MessageType)remains[(int)MessageFrame.Type].ConvertToInt32();
+            var peer = remains[(int)MessageFrame.Peer].ToByteArray();
+            byte[] signature = remains[(int)MessageFrame.Sign].ToByteArray();
+
+            NetMQFrame[] body = remains.Skip(CommonFrames).ToArray();
 
             // FIXME: The below code is too repetitive and prone to miss to add, which means,
             // when you add a new message type, you adds an enum member to MessageType and
@@ -157,6 +225,7 @@ namespace Libplanet.Net.Messages
                 { MessageType.ChainStatus, typeof(ChainStatus) },
                 { MessageType.GetBlockStates, typeof(GetBlockStates) },
                 { MessageType.BlockStates, typeof(BlockStates) },
+                { MessageType.DifferentVersion, typeof(DifferentVersion) },
             };
 
             if (!types.TryGetValue(rawType, out Type type))
@@ -167,7 +236,8 @@ namespace Libplanet.Net.Messages
 
             var message = (Message)Activator.CreateInstance(
                 type, new[] { body });
-            message.Remote = DeserializePeer(peer);
+            message.Version = remoteVersion;
+            message.Remote = remotePeer;
 
             if (!message.Remote.PublicKey.Verify(body.ToByteArray(), signature))
             {
@@ -184,7 +254,7 @@ namespace Libplanet.Net.Messages
             return message;
         }
 
-        public NetMQMessage ToNetMQMessage(PrivateKey key, Peer peer)
+        public NetMQMessage ToNetMQMessage(PrivateKey key, Peer peer, AppProtocolVersion version)
         {
             if (peer is null)
             {
@@ -203,6 +273,7 @@ namespace Libplanet.Net.Messages
             message.Push(key.Sign(message.ToByteArray()));
             message.Push(SerializePeer(peer));
             message.Push((byte)Type);
+            message.Push(version.Token);
 
             if (Identity is byte[] to)
             {
@@ -225,6 +296,32 @@ namespace Libplanet.Net.Messages
             using MemoryStream stream = new MemoryStream();
             formatter.Serialize(stream, peer);
             return stream.ToArray();
+        }
+
+        private static bool IsAppProtocolVersionValid(
+            Peer remotePeer,
+            AppProtocolVersion localVersion,
+            AppProtocolVersion remoteVersion,
+            IImmutableSet<PublicKey> trustedAppProtocolVersionSigners,
+            DifferentAppProtocolVersionEncountered differentAppProtocolVersionEncountered)
+        {
+            if (remoteVersion.Equals(localVersion))
+            {
+                return true;
+            }
+
+            if (!(trustedAppProtocolVersionSigners is null) &&
+                !trustedAppProtocolVersionSigners.Any(remoteVersion.Verify))
+            {
+                return false;
+            }
+
+            if (differentAppProtocolVersionEncountered is null)
+            {
+                return false;
+            }
+
+            return differentAppProtocolVersionEncountered(remotePeer, remoteVersion, localVersion);
         }
     }
 }